requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
bigquery==0.0.23
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==0.9.3
cryptography==38.0.3
dacktool==0.0.7
dbstream==0.1.15
DBUtils==3.0.2
Flask==2.2.2
frozenlist==1.3.3
google-api-core==2.8.2
google-api-python-client==1.7.11
google-auth==2.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.3.2
google-cloud-resource-manager==1.6.1
google-cloud-secret-manager==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
googleauthentication==0.0.16
greenlet==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==0.4.7
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
proto-plus==1.22.0
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.44
typing-inspect==0.8.0
typing_extensions==4.3.0
uritemplate==3.0.1
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
